ethics and data/igs-dashboard/requirements.txt
streamlit==1.38.0
pandas==2.2.2
plotly==5.24.1
requests==2.32.3
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.31
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-dotenv==1.0.1
nh3==0.2.18
pytest==8.3.2
python-multipart==0.0.6